Renaming modules: flixComps -> components flixCompsExperimental -> components_experimental flixBasics -> core flixStructure -> structure flixFeatures -> features basicModeling -> modeling
