+++ flixOpt/core.py
